ref: unit test restructure
